fix(scale-smoke): move the SQLite recovery scale test into kx-projection (off the llamacpp build path)
The M2.2b end-to-end SQLite recovery scale test lived in kx-runtime, which
transitively links kx-llamacpp(-sys); its cmake C++ build is not provisioned in
the lean `scale-smoke` CI job, so the job failed building the FFI (not a perf or
correctness failure — the two kx-projection scale cases passed: ratio 1.00,
speedup 2.5x).

Move the identical SQLite-backed proof into kx-projection/tests/fold_checkpoint.rs
(it owns FoldCheckpoint and reaches kx-journal::SqliteJournal — rusqlite, no
llamacpp), folding via Projection::from_journal[_with_checkpoint] over a real
disk-backed journal + an on-disk sidecar blob. Drop the kx-runtime line from the
scale-smoke recipe. The kx-runtime checkpoint_io atomic-sidecar wiring stays
covered by kx-runtime's own `test` job (checkpoint_recovery + checkpoint_io units
+ kill_and_replay). Local: 3.9x speedup (105k entries / 5k live Motes).
